HH-60953 rename packages to be consist with artifact ids [i.garanina]
