+++ httpemulator-client/src/main/java/ru/hh/httpemulator/client/entity/RestrictionType.java
@@ -1,0 +1,24 @@
+package ru.hh.httpemulator.client.entity;
+
+public enum RestrictionType {
+
+	EQ(1),
+	
+	OR(101),
+	
+	AND(102);
+	
+	private final int value;
+
+	private RestrictionType(int value) {
+		this.value = value;
+	}
+
+	public int getValue() {
+		return value;
+	}
+	
+	public boolean isAgregate(){
+		return value > 100;
+	}
+}
