--- httpemulator-server/src/main/java/ru/hh/http/emulator/exception/AmbiguousRulesException.java
@@ -1,14 +1,0 @@
-package ru.hh.http.emulator.exception;
-
-public class AmbiguousRulesException extends Exception {
-
-	public AmbiguousRulesException(String message, Throwable cause) {
-		super(message, cause);
-	}
-
-	public AmbiguousRulesException(String message) {
-		super(message);
-	}
-
-	
-}
