--- httpemulator-server/src/main/java/ru/hh/http/emulator/exception/RuleNotFoundException.java
@@ -1,13 +1,0 @@
-package ru.hh.http.emulator.exception;
-
-public class RuleNotFoundException extends Exception {
-
-	public RuleNotFoundException(String message, Throwable cause) {
-		super(message, cause);
-	}
-
-	public RuleNotFoundException(String message) {
-		super(message);
-	}
-	
-}
