--- httpemulator-server/src/main/java/ru/hh/http/emulator/exception/ScenarioNotFoundException.java
@@ -1,11 +1,0 @@
-package ru.hh.http.emulator.exception;
-
-public class ScenarioNotFoundException extends Exception {
-
-	public ScenarioNotFoundException(String message) {
-		super(message);
-	}
-
-	
-	
-}
