--- httpemulator-server/src/main/java/ru/hh/http/emulator/utils/CharsetUtils.java
@@ -1,9 +1,0 @@
-package ru.hh.http.emulator.utils;
-
-import java.nio.charset.Charset;
-
-public class CharsetUtils {
-
-	public static final Charset UTF_8 = Charset.forName("UTF-8");
-	
-}
