+++ httpemulator-server/src/main/java/ru/hh/httpemulator/server/exception/RuleNotFoundException.java
@@ -1,0 +1,13 @@
+package ru.hh.httpemulator.server.exception;
+
+public class RuleNotFoundException extends Exception {
+
+	public RuleNotFoundException(String message, Throwable cause) {
+		super(message, cause);
+	}
+
+	public RuleNotFoundException(String message) {
+		super(message);
+	}
+	
+}
