+++ httpemulator-server/src/main/java/ru/hh/httpemulator/server/exception/ScenarioNotFoundException.java
@@ -1,0 +1,11 @@
+package ru.hh.httpemulator.server.exception;
+
+public class ScenarioNotFoundException extends Exception {
+
+	public ScenarioNotFoundException(String message) {
+		super(message);
+	}
+
+	
+	
+}
