+++ httpemulator-server/src/main/java/ru/hh/httpemulator/server/utils/CharsetUtils.java
@@ -1,0 +1,9 @@
+package ru.hh.httpemulator.server.utils;
+
+import java.nio.charset.Charset;
+
+public class CharsetUtils {
+
+	public static final Charset UTF_8 = Charset.forName("UTF-8");
+	
+}
