+++ httpemulator-server/src/test/resources/test-context.xml
@@ -1,8 +1,8 @@
 <beans xmlns="http://www.springframework.org/schema/beans"
-	xmlns:context="http://www.springframework.org/schema/context"
-	xmlns:tx="http://www.springframework.org/schema/tx"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="
+       xmlns:context="http://www.springframework.org/schema/context"
+       xmlns:tx="http://www.springframework.org/schema/tx"
+       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+       xsi:schemaLocation="
         http://www.springframework.org/schema/beans     
         http://www.springframework.org/schema/beans/spring-beans-3.0.xsd
         http://www.springframework.org/schema/context 
@@ -10,11 +10,11 @@
         http://www.springframework.org/schema/tx
      	http://www.springframework.org/schema/tx/spring-tx-3.0.xsd">
 
-	<context:component-scan base-package="ru.hh.http.emulator" />
+    <context:component-scan base-package="ru.hh.httpemulator" />
   
-	<bean id="objectMapper" class="com.fasterxml.jackson.databind.ObjectMapper"/>
+    <bean id="objectMapper" class="com.fasterxml.jackson.databind.ObjectMapper"/>
   
-	<bean id="criteriaEngine" class="ru.hh.http.emulator.engine.CriteriaHttpEngine">
-    	<property name="expirationTimeout" value="36000"/>
-	</bean>
+    <bean id="criteriaEngine" class="ru.hh.httpemulator.server.engine.CriteriaHttpEngine">
+        <property name="expirationTimeout" value="36000"/>
+    </bean>
 </beans>
